pyproject.toml
[tool.poetry]
name = "amazon-api"
version = "0.1.0"
description = "Django + DRF Scraping Amazon Site"
authors = ["Feri Lukmansyah <47948861+perymerdeka@users.noreply.github.com>"]
readme = "README.md"

# ...
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
Django = "^4.2.6"
djangorestframework = "^3.14.0"
httpx = "^0.25.0"
beautifulsoup4 = "^4.12.2"
fake-useragent = "^1.3.0"
loguru = "^0.7.2"
rich = "^13.6.0"
selectolax = "^0.3.17"
backoff = "^2.2.1"
pandas = "^2.1.2"
openpyxl = "^3.1.2"
requests = "^2.31.0"
playwright = "^1.39.0"
django-allauth = "^0.58.1"
djangorestframework-simplejwt = "^5.3.0"
setuptools = "^68.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
